pyproject.toml
[tool.poetry]
name = "aio-evergreen.py"
version = "0.1.1"
description = "Async Evergreen API client for python"
authors = ["David Bradford <david.bradford@mongodb.com>"]
license = "Apache-2.0"
readme = "README.md"
repository = "https://github.com/dbradf/aio-evergreen.py"
packages = [
    {include = "evg", from = "src"},
]

[tool.poetry.scripts]
evg-api = 'evg.cli.cli:main'

[tool.poetry.dependencies]
python = "^3.6"
pydantic = "^1"
aiohttp = "^3"
pyyaml = "^5"

[tool.poetry.dev-dependencies]
pytest = "^6"
pytest-black = "^0.3"
pytest-cov = "^2.8"
pytest-flake8 = "^1.0"
pytest-mypy = "^0.4"
pytest-pydocstyle = "^2.0"
pytest-sugar = "^0.9"
pytest-isort = "^0.3"
black = "^20.8b1"

[tool.black]
line-length = 100
target-version = ['py36']
